changed structure to reflect various middlewares
